[#76] Better show('spect_fit')
